+++ frictionless/fields/boolean.py
@@ -1,8 +1,9 @@
+from __future__ import annotations
 from ..schema.field import Field
-
 
 class BooleanField(Field):
     ### TEMP Only required for Metadata compatibility
     ### This is required because "metadata_import" makes a distinction based
     ### on the "type" property (`is_typed_class`)
     type = "boolean"
+
