+++ frictionless/fields/date.py
@@ -1,9 +1,7 @@
 from __future__ import annotations
 
 import attrs
-
 from ..schema import Field
-
 
 @attrs.define(kw_only=True, repr=False)
 class DateField(Field):
@@ -15,3 +13,4 @@ class DateField(Field):
         "maximum",
         "enum",
     ]
+
