+++ frictionless/fields/field_descriptor.py
@@ -1,16 +1,11 @@
-"""field_descriptor.py provides pydantic Models for Field descriptors"""
-
 from __future__ import annotations
 
 import datetime
-from typing import Any, ClassVar, Dict, List, Literal, Optional, Union, Pattern
-import re
+from typing import List, Literal, Optional, Union
 
-from pydantic import Field as PydanticField, AliasChoices, model_validator, BaseModel
-from typing_extensions import Self
+from pydantic import Field as PydanticField, BaseModel
 
-
-from .. import settings
+from .base_field_descriptor import BaseFieldDescriptor
 from .field_constraints import (
     BaseConstraints,
     CollectionConstraints,
@@ -18,125 +13,6 @@ from .field_constraints import (
     StringConstraints,
     ValueConstraints,
 )
-
-TableSchemaTypes = Union[bool, str, float, int]
-"""Python equivalents of types supported by the Table schema specification"""
-
-
-class BaseFieldDescriptor(BaseModel):
-    """Data model of a (unspecialised) field descriptor"""
-
-    name: str
-    """
-    The field descriptor MUST contain a name property.
-    """
-
-    title: Optional[str] = None
-    """
-    A human readable label or title for the field
-    """
-
-    description: Optional[str] = None
-    """
-    A description for this field e.g. “The recipient of the funds”
-    """
-
-    missing_values: Optional[List[str]] = PydanticField(
-        default=None, alias="missingValues"
-    )
-    """
-    A list of field values to consider as null values
-    """
-
-    example: Optional[Any] = None
-    """
-    An example of a value for the field.
-    """
-
-    @model_validator(mode="before")
-    @classmethod
-    def compat(cls, data: Dict[str, Any]) -> Dict[str, Any]:
-        # Backward compatibility for field.format
-
-        format_ = data.get("format")
-        if format_:
-            if format_.startswith("fmt:"):
-                data["format"] = format_[4:]
-
-        return data
-
-    @model_validator(mode="after")
-    def validate_example(self) -> Self:
-        """Validate that the example value can be converted using read_value() if available"""
-        if self.example is not None:
-            if hasattr(self, "read_value"):
-                read_value_method = getattr(self, "read_value")
-                result = read_value_method(self.example)
-                if result is None:
-                    raise ValueError(
-                        f'example value "{self.example}" for field "{self.name}" is not valid'
-                    )
-
-        return self
-
-
-class BooleanFieldDescriptor(BaseFieldDescriptor):
-    """The field contains boolean (true/false) data."""
-
-    type: ClassVar[Literal["boolean"]] = "boolean"
-
-    format: Optional[Literal["default"]] = None
-    constraints: Optional[BaseConstraints[bool]] = None
-
-    true_values: Optional[List[str]] = PydanticField(
-        default=settings.DEFAULT_TRUE_VALUES,
-        alias="trueValues",
-        validation_alias=AliasChoices("trueValues", "true_values"),
-    )
-    """
-    Values to be interpreted as “true” for boolean fields
-    """
-
-    false_values: Optional[List[str]] = PydanticField(
-        default=settings.DEFAULT_FALSE_VALUES,
-        alias="falseValues",
-        validation_alias=AliasChoices("falseValues", "false_values"),
-    )
-    """
-    Values to be interpreted as “false” for boolean fields
-    """
-
-    def read_value(self, cell: TableSchemaTypes) -> Optional[bool]:
-        """read_value converts the physical (possibly typed) representation to
-        a logical boolean representation.
-
-        See "Data representation" in the glossary for more details.
-        https://datapackage.org/standard/glossary/#data-representation
-
-        If the physical representation is already typed as a boolean, the
-        value is returned unchanged.
-
-        If the physical representation is a string, then the string is parsed
-        as a boolean depending on true_values and false_values options. `None`
-        is returned if the string cannot be parsed.
-
-        Any other typed input will return `None`.
-        """
-        if isinstance(cell, bool):
-            return cell
-
-        if isinstance(cell, str):
-            if self.true_values and cell in self.true_values:
-                return True
-            if self.false_values and cell in self.false_values:
-                return False
-
-        return None
-
-    def write_value(self, cell: Optional[bool]) -> Optional[str]:
-        if self.true_values and self.false_values:
-            return self.true_values[0] if cell else self.false_values[0]
-        return None
 
 
 class ArrayFieldDescriptor(BaseFieldDescriptor):
@@ -146,8 +22,9 @@ class ArrayFieldDescriptor(BaseFieldDescriptor):
     format: Optional[Literal["default"]] = None
     constraints: Optional[JSONConstraints] = None
 
-    # TODO type is not accurate : array item are unnamed, not described etc
-    array_item: Optional[FieldDescriptor] = PydanticField(
+    # TODO type is not accurate : array item are unnamed, not described etc
+    # Using string annotation to avoid circular import
+    array_item: Optional["FieldDescriptor"] = PydanticField(
         default=None, alias="arrayItem"
     )
 
@@ -158,52 +35,6 @@ class AnyFieldDescriptor(BaseFieldDescriptor):
     type: Literal["any"] = "any"
     format: Optional[Literal["default"]] = None
     constraints: Optional[BaseConstraints[str]] = None
-
-
-class DateFieldDescriptor(BaseFieldDescriptor):
-    """The field contains a date without a time."""
-
-    type: Literal["date"] = "date"
-    format: Optional[str] = None
-    constraints: Optional[ValueConstraints[str]] = None
-
-    def read_value(self, cell: Any) -> Optional[datetime.date]:
-        from datetime import date, datetime
-        from ..platform import platform
-
-        if isinstance(cell, datetime):
-            value_time = cell.time()
-            if (
-                value_time.hour == 0
-                and value_time.minute == 0
-                and value_time.second == 0
-            ):
-                return datetime(cell.year, cell.month, cell.day).date()
-            else:
-                return None
-        if isinstance(cell, date):
-            return cell
-        if not isinstance(cell, str):
-            return None
-        try:
-            format_value = self.format or "default"
-            if format_value == "default":
-                cell = datetime.strptime(cell, settings.DEFAULT_DATE_PATTERN).date()
-            elif format_value == "any":
-                cell = platform.dateutil_parser.parse(cell).date()
-            else:
-                cell = datetime.strptime(cell, format_value).date()
-        except Exception:
-            return None
-        return cell
-
-    def write_value(self, cell: Optional[datetime.date]) -> Optional[str]:
-        if cell is None:
-            return None
-        format_value = self.format or "default"
-        if format_value == settings.DEFAULT_FIELD_FORMAT:
-            format_value = settings.DEFAULT_DATE_PATTERN
-        return cell.strftime(format_value)
 
 
 class DatetimeFieldDescriptor(BaseFieldDescriptor):
@@ -245,6 +76,7 @@ class GeoPointFieldDescriptor(BaseFieldDescriptor):
 
 
 class CategoryDict(BaseModel):
+    """Category dictionary for field categories."""
     value: str
     label: Optional[str] = None
 
@@ -253,98 +85,7 @@ ICategories = Union[
     List[str],
     List[CategoryDict],
 ]
-
-
-class IntegerFieldDescriptor(BaseFieldDescriptor):
-    """The field contains integers - that is whole numbers."""
-
-    type: Literal["integer"] = "integer"
-    format: Optional[Literal["default"]] = None
-    constraints: Optional[ValueConstraints[int]] = None
-
-    categories: Optional[ICategories] = None
-    """
-    Property to restrict the field to a finite set of possible values
-    """
-
-    categories_ordered: Optional[bool] = PydanticField(
-        default=None, alias="categoriesOrdered"
-    )
-    """
-    When categoriesOrdered is true, implementations SHOULD regard the order of
-    appearance of the values in the categories property as their natural order.
-    """
-
-    group_char: Optional[str] = PydanticField(default=None, alias="groupChar")
-    """
-    String whose value is used to group digits for integer/number fields
-    """
-
-    bare_number: bool = PydanticField(
-        default=settings.DEFAULT_BARE_NUMBER, alias="bareNumber"
-    )
-    """
-    If false leading and trailing non numbers will be removed for integer/number fields
-    """
-
-    pattern: ClassVar[Pattern[str]] = re.compile(r"((^[^-\d]*)|(\D*$))")
-
-    def read_value(self, cell: Any) -> Optional[int]:
-        """read_value converts the physical (possibly typed) representation to
-        a logical integer representation.
-
-        See "Data representation" in the glossary for more details.
-        https://datapackage.org/standard/glossary/#data-representation
-
-        If the physical representation is already typed as an integer, the
-        value is returned unchanged.
-
-        If the physical representation is a string, then the string is parsed
-        as an integer. If `bare_number` is False, non-digit characters are
-        removed first. `None` is returned if the string cannot be parsed.
-
-        If the physical representation is a float or Decimal that represents
-        a whole number, it is converted to an integer.
-
-        Any other typed input will return `None`.
-        """
-        from decimal import Decimal
-
-        if isinstance(cell, bool):
-            return None
-
-        elif isinstance(cell, int):
-            return cell
-
-        elif isinstance(cell, str):
-            cell = cell.strip()
-
-            # Process the cell (remove non-digit characters if bare_number is False)
-            if not self.bare_number:
-                cell = self.pattern.sub("", cell)
-
-            # Cast the cell
-            try:
-                return int(cell)
-            except Exception:
-                return None
-
-        elif isinstance(cell, float) and cell.is_integer():
-            return int(cell)
-        elif isinstance(cell, Decimal) and cell % 1 == 0:
-            return int(cell)
-
-        return None
-
-    def write_value(self, cell: Optional[int]) -> Optional[str]:
-        """write_value converts the logical integer representation to
-        a physical (string) representation.
-
-        Returns the integer as a string.
-        """
-        if cell is None:
-            return None
-        return str(cell)
+"""Categories type used by IntegerFieldDescriptor and StringFieldDescriptor"""
 
 
 IItemType = Literal[
@@ -464,17 +205,16 @@ class YearmonthFieldDescriptor(BaseFieldDescriptor):
     format: Optional[Literal["default"]] = None
     constraints: Optional[ValueConstraints[str]] = None
 
-
 FieldDescriptor = Union[
     AnyFieldDescriptor,
-    ArrayFieldDescriptor,
-    BooleanFieldDescriptor,
-    DateFieldDescriptor,
+    ArrayFieldDescriptor,  # wip
+    # BooleanFieldDescriptor,  # v
+    # DateFieldDescriptor,  # v
     DatetimeFieldDescriptor,
     DurationFieldDescriptor,
     GeoJSONFieldDescriptor,
     GeoPointFieldDescriptor,
-    IntegerFieldDescriptor,
+    # IntegerFieldDescriptor,  # v
     ListFieldDescriptor,
     NumberFieldDescriptor,
     ObjectFieldDescriptor,
